Added basic versions of search command.
If provided no name, search command will list all of a users existing
playlists. If provided a name, it will search through a user's existing
playlists to try and find a match.

If finds it, it returns it.
If it does not find it, it tells the user.

Currently, I do nothing with the returned playlist.
I could display more info about the playlist.

I also want to make the search more robust, try to return similar items.