Wrote new tests. Bypassing pytest.
Wrote new tests to describe functionality related to...
The fact that spotify allows multiple playlists with the same name.

Some code needs to be refactored as a result of that fact, and the
tests are the start of that refactoring.

They are failing, as they should right now.

I will write the code to pass them tomorrow.

As such, I'm bypassing the auto-testing on pre-commit.