Refactor get_item methods.
Since get_item() was getting passed a Spotify instance anyways,
might as well make it static.

Additionally, define a base, static version of get_item() in Item that
sub-classes can invoke and give a concrete getter to.

Benefit of doing this is so that the exception handling code is not
duplicated across many subclasses of item.